Initial implementation of token refresh

+++ pkg/api/handlers/oauth/authresponses.go
@@ -8,8 +8,8 @@ type tokenResponse struct {
 	State        string `json:"state"`
 	TokenType    string `json:"token_type"`
 	Scope        string `json:"scope"`
-	ExpiresIn    int    `json:"expires_in"`
-	ExtExpiresIn int    `json:"ext_expires_in"`
+	ExpiresIn    int64  `json:"expires_in"`
+	ExtExpiresIn int64  `json:"ext_expires_in"`
 	AccessToken  string `json:"access_token"`
 	RefreshToken string `json:"refresh_token"`
 	Ok           bool   `json:"ok"`
